Handle Directory removal error uniformly
- Based on OS the Directory removal may return EPERM instead of EISDIR error. This fix checks if we are removing a Directory and will ensure EISDIR return even when on MacOS

+++ localfs.js
@@ -621,7 +621,21 @@ module.exports = function setup(fsOptions) {
     }
 
     function rmfile(path, options, callback) {
-        remove(path, fs.unlink, callback);
+      resolvePath(path, function (err, realpath) {
+        if (err) return callback(err);
+
+        fs.lstat(realpath, function (err, stat) {
+          if (err) return callback(err);
+
+          if (stat.isDirectory()) {
+            const e = new Error("EISDIR: Requested resource is a directory");
+            e.code = "EISDIR";
+            return callback(e);
+          }
+
+          remove(path, fs.unlink, callback);
+        });
+      });
     }
 
     function rmdir(path, options, callback) {
